requirements.txt
streamlit==1.37.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pandas==2.2.2
requests==2.31.0
python-dotenv==1.0.1
numpy==1.26.4